# api/openapi.yaml
openapi: 3.1.0
info:
  title: Zoomies API
  version: "1.0.0"
  summary: Self-hosted GitHub Actions runner fleet controller
  description: |
    Zoomies manages ephemeral self-hosted GitHub Actions runners.

    This document is the whole product surface. The web UI and the `zoomies`
    CLI are both clients of it, and nothing is reachable from either that is
    not described here.

    ## Authentication

    Two credentials are accepted on the same routes:

    * a **session cookie** (`zoomies_session`), which the browser gets from
      `POST /auth/login`;
    * a **bearer token** (`Authorization: Bearer zoo_...`), for automation.

    Cookie-authenticated mutating requests are subject to a same-origin check.
    Bearer-authenticated requests are not, because they are not vulnerable to
    CSRF.

    Agents use a separate credential class on `/agent/*` and can reach nothing
    else.

    ## Roles

    | Role | May |
    |---|---|
    | `viewer` | read everything except secrets |
    | `operator` | act on the fleet and manage pools |
    | `admin` | manage users, tokens, installations and settings |

    Each operation below records its minimum role in `x-zoomies-role`.

    ## Conventions

    * List endpoints take `limit`, `offset`, `sort` and `order`, and return
      `{ items, total, limit, offset }`.
    * Timestamps are RFC 3339 with a `Z` offset.
    * Durations are Go duration strings: `"5m"`, `"1h30s"`.
...
      message written for a human to read.
  license:
...
servers:
...
security:
  - cookieAuth: []
  - bearerAuth: []
...
paths:
...
  /pools/{id}:
    parameters: [{ $ref: "#/components/parameters/PathID" }]
    get:
      operationId: getPool
      tags: [pools]
      summary: Get a pool
      x-zoomies-role: viewer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Pool" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updatePool
      tags: [pools]
      summary: Update a pool
      x-zoomies-role: operator
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PoolUpdate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Pool" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/Unprocessable" }
    delete:
      operationId: deletePool
      tags: [pools]
      summary: Delete a pool
      x-zoomies-role: operator
      parameters:
        - name: drain
          in: query
          description: Let running jobs finish first. The default, and the only safe choice while work is in flight.
          schema: { type: boolean, default: true }
        - name: force
          in: query
          description: Destroy runners immediately, interrupting any job they are running.
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  runners_affected: { type: integer }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /runners:
    get:
      operationId: listRunners
      tags: [runners]
      summary: List runners
      x-zoomies-role: viewer
      parameters:
        - { name: pool_id, in: query, schema: { type: array, items: { type: string } }, style: form, explode: true }
        - { name: host_id, in: query, schema: { type: array, items: { type: string } }, style: form, explode: true }
        - name: state
          in: query
          schema: { type: array, items: { $ref: "#/components/schemas/RunnerState" } }
          style: form
          explode: true
        - { name: q, in: query, description: Substring match on name, ID or container ID., schema: { type: string } }
        - name: include_removed
          in: query
          description: Terminal runners are hidden by default, because a busy fleet accumulates thousands of them.
          schema: { type: boolean, default: false }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/Order"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/Runner" }
  /runners/{id}:
    parameters: [{ $ref: "#/components/parameters/PathID" }]
    get:
      operationId: getRunner
      tags: [runners]
      summary: Get a runner
      x-zoomies-role: viewer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RunnerDetail" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteRunner
      tags: [runners]
      summary: Delete a runner
      description: Without `force` this drains first, so a running job is never interrupted. The registration is removed from GitHub either way.
      x-zoomies-role: operator
      parameters:
        - name: force
          in: query
          schema: { type: boolean, default: false }
      responses:
        "202": { description: Accepted }
        "404": { $ref: "#/components/responses/NotFound" }
  /runners/{id}/drain:
    parameters: [{ $ref: "#/components/parameters/PathID" }]
    post:
      operationId: drainRunner
      tags: [runners]
      summary: Drain a runner
      description: Finish the current job, then exit. Never interrupts work in progress.
      x-zoomies-role: operator
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Runner" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: The runner is already terminal.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  /hosts/{id}:
    parameters: [{ $ref: "#/components/parameters/PathID" }]
    get:
      operationId: getHost
      tags: [hosts]
      summary: Get a host
      x-zoomies-role: viewer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Host" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateHost
      tags: [hosts]
      summary: Update a host's capacity or labels
      x-zoomies-role: operator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                capacity: { type: integer, minimum: 0 }
                labels: { type: object, additionalProperties: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Host" }
    delete:
      operationId: deleteHost
      tags: [hosts]
      summary: Remove a host
      description: Refused while the host still has live runners, unless forced.
      x-zoomies-role: admin
      parameters:
        - { name: force, in: query, schema: { type: boolean, default: false } }
      responses:
        "204": { description: Removed }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: zoomies_session
      description: Set by `POST /auth/login`. HttpOnly, SameSite=Lax, and Secure when the external URL is https.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: zoo_<prefix>_<secret>
      description: An API token. Not subject to the same-origin check, because it is not vulnerable to CSRF.
  parameters:
    PathID:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: The resource ID, e.g. `pool_k3f9qz2m`.
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 50, minimum: 1, maximum: 500 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, default: 0, minimum: 0 }
    Sort:
      name: sort
      in: query
      description: A column name. An unknown value falls back to the default rather than erroring, so a stale bookmark does not break the page.
      schema: { type: string }
    Order:
      name: order
      in: query
      schema: { type: string, enum: [asc, desc], default: desc }
  responses:
...
    NotFound:
      description: No such thing.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Conflict:
      description: The request conflicts with the current state.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Unprocessable:
      description: The request was understood but is not valid. `errors` names the fields.
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ErrorEnvelope"
              - type: object
                properties:
                  errors: { type: array, items: { $ref: "#/components/schemas/FieldError" } }
...
  schemas:
...
    BackendKind:
      type: string
      enum: [docker, podman, process]
    DockerMode:
      type: string
      enum: [none, dind, host-socket]
      description: |
        How much Docker a job on this pool gets.
        `none` is the default and the only one that is not a warning.
        `dind` gives a private, privileged sidecar daemon.
        `host-socket` hands the job the host's daemon, and with it root on the host.
    RunnerState:
      type: string
      enum: [provisioning, registering, idle, busy, draining, removed, failed]
    JobState:
      type: string
      enum: [queued, in_progress, completed]
    Severity:
      type: string
      enum: [error, warning, info]
    Duration:
      type: string
      description: A Go duration string.
      examples: ["5m", "1h30s"]
    Page:
      type: object
      required: [total, limit, offset]
      properties:
        total: { type: integer, description: Rows matching the filter, not rows returned. }
        limit: { type: integer }
        offset: { type: integer }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [bad_request, unauthorized, forbidden, not_found, conflict, unprocessable, rate_limited, internal]
            message: { type: string, description: Written for a person to read, and to act on. }
            field: { type: string }
            detail: { type: string }
    FieldError:
      type: object
      required: [field, message]
      properties:
        field: { type: string, examples: [labels] }
        message: { type: string, examples: ["a pool needs at least one label your workflows can ask for"] }
    Problem:
      type: object
      required: [code, severity, title]
      properties:
        code: { type: string, examples: [bind.public_no_tls] }
        severity: { $ref: "#/components/schemas/Severity" }
        setting: { type: string, examples: [server.bind] }
        title: { type: string, description: One line - what is true. }
        detail: { type: string, description: Why it matters, and the consequence. }
        fix: { type: string, description: What to change, concretely. }
        target_kind: { type: string }
        target_id: { type: string }
        since: { type: string, format: date-time }

    # --- meta -------------------------------------------------------------
...
    Resources:
      type: object
      properties:
        cpus: { type: number, format: double, examples: [2] }
        memory_mb: { type: integer, format: int64, examples: [4096] }
        disk_gb: { type: integer, format: int64 }
        pids_limit: { type: integer, format: int64 }
    Pool:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        installation_id: { type: string }
        installation_target: { type: string }
        labels: { type: array, items: { type: string } }
        runner_group: { type: string }
        backend: { $ref: "#/components/schemas/BackendKind" }
        image: { type: string }
        runner_version: { type: string }
        min_runners: { type: integer }
        max_runners: { type: integer }
        idle_timeout: { $ref: "#/components/schemas/Duration" }
        ephemeral: { type: boolean }
        docker_mode: { $ref: "#/components/schemas/DockerMode" }
        resources: { $ref: "#/components/schemas/Resources" }
        host_selector: { type: object, additionalProperties: { type: string } }
        env: { type: object, additionalProperties: { type: string } }
        run_as_root: { type: boolean }
        enabled: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        counts:
          type: object
          properties:
            provisioning: { type: integer }
            registering: { type: integer }
            idle: { type: integer }
            busy: { type: integer }
            draining: { type: integer }
            failed: { type: integer }
            live: { type: integer }
        queued_jobs: { type: integer }
        utilisation: { type: number, format: double }
        warnings:
          type: array
          items: { $ref: "#/components/schemas/Problem" }
          description: The dangerous settings this pool has in effect, if any.
    PoolCreate:
      type: object
      required: [name, installation_id, labels, backend]
      properties:
        name: { type: string, minLength: 1, maxLength: 64, examples: [linux-x64] }
        installation_id: { type: string }
        labels: { type: array, items: { type: string }, minItems: 1 }
        runner_group: { type: string }
        backend: { $ref: "#/components/schemas/BackendKind" }
        image: { type: string }
        runner_version: { type: string }
        min_runners: { type: integer, default: 0, minimum: 0 }
        max_runners: { type: integer, default: 4, minimum: 1 }
        idle_timeout: { allOf: [{ $ref: "#/components/schemas/Duration" }], default: "5m" }
        ephemeral: { type: boolean, default: true }
        docker_mode: { allOf: [{ $ref: "#/components/schemas/DockerMode" }], default: none }
        resources: { $ref: "#/components/schemas/Resources" }
        host_selector: { type: object, additionalProperties: { type: string } }
        env: { type: object, additionalProperties: { type: string } }
        run_as_root: { type: boolean, default: false }
        enabled: { type: boolean, default: true }
    PoolUpdate:
      allOf:
        - $ref: "#/components/schemas/PoolCreate"
        - type: object
          required: []

    # --- runners ----------------------------------------------------------
    Runner:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        pool_id: { type: string }
        pool_name: { type: string }
        host_id: { type: string }
        host_name: { type: string }
        state: { $ref: "#/components/schemas/RunnerState" }
        github_runner_id: { type: integer, format: int64 }
        container_id: { type: string }
        ephemeral: { type: boolean }
        labels: { type: array, items: { type: string } }
        image: { type: string }
        runner_version: { type: string }
        current_job_id: { type: string }
        current_job: { $ref: "#/components/schemas/Job" }
        message: { type: string }
        jobs_handled: { type: integer }
        cpu_percent: { type: number, format: double }
        memory_bytes: { type: integer, format: int64 }
        created_at: { type: string, format: date-time }
        started_at: { type: [string, "null"], format: date-time }
        last_idle_at: { type: [string, "null"], format: date-time }
        finished_at: { type: [string, "null"], format: date-time }
    RunnerDetail:
      allOf:
        - $ref: "#/components/schemas/Runner"
        - type: object
          properties:
            host: { $ref: "#/components/schemas/Host" }
            pool: { $ref: "#/components/schemas/Pool" }
            timeline: { type: array, items: { $ref: "#/components/schemas/TimelineEntry" } }
            logs_available: { type: boolean, description: False when the runner's host is unreachable, so the UI can say why rather than showing an empty log. }
    TimelineEntry:
      type: object
      properties:
        state: { $ref: "#/components/schemas/RunnerState" }
        at: { type: string, format: date-time }
        duration_ms: { type: integer, format: int64, description: How long the runner stayed in this state. }
        message: { type: string }

    # --- jobs -------------------------------------------------------------
    Job:
      type: object
      properties:
        id: { type: string }
        github_job_id: { type: integer, format: int64 }
        github_run_id: { type: integer, format: int64 }
        repo: { type: string }
        workflow: { type: string }
        job_name: { type: string }
        labels: { type: array, items: { type: string } }
        state: { $ref: "#/components/schemas/JobState" }
        conclusion: { type: string, examples: [success, failure, cancelled, skipped] }
        pool_id: { type: string }
        pool_name: { type: string }
        runner_id: { type: string }
        runner_name: { type: string }
        html_url: { type: string, description: The run on GitHub. }
        matched: { type: boolean, description: False when no enabled pool claims this job's labels. }
        queued_at: { type: string, format: date-time }
        started_at: { type: [string, "null"], format: date-time }
        completed_at: { type: [string, "null"], format: date-time }
        queue_wait_ms: { type: integer, format: int64 }
        duration_ms: { type: integer, format: int64 }

    # --- hosts ------------------------------------------------------------
    BackendInfo:
      type: object
      properties:
        kind: { $ref: "#/components/schemas/BackendKind" }
        available: { type: boolean }
        version: { type: string }
        rootless: { type: boolean }
        endpoint: { type: string }
        detail: { type: string, description: When unavailable, what is wrong and how to fix it. }
        supports_dind: { type: boolean }
    Host:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        address: { type: string }
        embedded: { type: boolean, description: True for the agent running inside the controller process. }
        capacity: { type: integer }
        active_runners: { type: integer }
        free: { type: integer }
        backends: { type: array, items: { type: string } }
        backend_info: { type: array, items: { $ref: "#/components/schemas/BackendInfo" } }
        labels: { type: object, additionalProperties: { type: string } }
        os: { type: string }
        arch: { type: string }
        version: { type: string }
        cordoned: { type: boolean }
        healthy: { type: boolean }
        last_heartbeat: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
...